Only spawn if not exists
This is the first step to making what I am calling "hubs-in-waiting".
You'll be able to have a main services server and multiple sasl-servers attached to each server, so you never have to worry about services going down affecting authentication.

In such a case of services going down, one of the sasl-servers will become services and respawn the clients.