Add continuous integration using irctest
On each commit/PR, this will make GitHub build UnrealIRCd,
and install it along with Wordpress, wp-cli, and irctest.

Then, it will run irctest's test suite, which tests the IRC server
answers to service-related commands as expected.
If it does not, it will report the issue on the GitHub UI next to the
commit/PR.