Extend MAIL, add user_list_by_account()
Users who are logged in with an account who is in receipt of a `/MAIL`, but is not directly logged in with that nick, will receive a notification letting them know they have mail and how to view it (`/MAIL -list`)

Added `user_list_by_account(String $account)`
Similar to `user_list`, except it queries the database only for users who are logged in with that account. This is quicker than cycling through all users. This is used by `MAIL`.

Other: Cleaned up some curly bracket placement to look cleaner.